Introduce abstract SteamWorkshopTask
